feat(driver,iour): move is_op_supported
