+++ compio-driver/src/fusion/mod.rs
@@ -11,6 +11,7 @@ pub use std::os::fd::{AsFd, AsRawFd, BorrowedFd, OwnedFd, RawFd};
 use std::{io, task::Poll, time::Duration};
 
 use compio_log::warn;
+pub(crate) use iour::is_op_supported;
 pub use iour::{OpCode as IourOpCode, OpEntry};
 pub use poll::{Decision, OpCode as PollOpCode, OpType};
