+++ compio-driver/src/iour/mod.rs
@@ -31,6 +31,29 @@ use crate::{AsyncifyPool, BufferPool, DriverType, Entry, Key, ProactorBuilder, s
 
 pub(crate) mod op;
 
+pub(crate) fn is_op_supported(code: u8) -> bool {
+    #[cfg(feature = "once_cell_try")]
+    use std::sync::OnceLock;
+
+    #[cfg(not(feature = "once_cell_try"))]
+    use once_cell::sync::OnceCell as OnceLock;
+
+    static PROBE: OnceLock<io_uring::Probe> = OnceLock::new();
+
+    PROBE
+        .get_or_try_init(|| {
+            let mut probe = io_uring::Probe::new();
+
+            io_uring::IoUring::new(2)?
+                .submitter()
+                .register_probe(&mut probe)?;
+
+            std::io::Result::Ok(probe)
+        })
+        .map(|probe| probe.is_supported(code))
+        .unwrap_or_default()
+}
+
 /// The created entry of [`OpCode`].
 pub enum OpEntry {
     /// This operation creates an io-uring submission entry.
